+++ packages/legend-application-repl/src/stores/dataCube/extend/DataCubeExtendManagerState.tsx
@@ -20,7 +20,7 @@ import type {
   DataCubeQuerySnapshotColumn,
   DataCubeQuerySnapshotExtendedColumn,
 } from '../core/DataCubeQuerySnapshot.js';
-import { guaranteeNonNullable } from '@finos/legend-shared';
+import { deleteEntry, guaranteeNonNullable, noop } from '@finos/legend-shared';
 import type { DataCubeState } from '../DataCubeState.js';
 import { DataCubeQuerySnapshotController } from '../core/DataCubeQuerySnapshotManager.js';
 import {
@@ -98,6 +98,7 @@ export class DataCubeExtendManagerState extends DataCubeQuerySnapshotController 
     column: T,
     isGroupLevel: boolean,
     columnKind: DataCubeColumnKind | undefined,
+    editor: DataCubeNewColumnState,
   ) {
     (isGroupLevel ? this.groupExtendedColumns : this.leafExtendedColumns).push(
       new DataCubeQueryExtendedColumnState(column),
@@ -111,6 +112,7 @@ export class DataCubeExtendManagerState extends DataCubeQuerySnapshotController 
     }
 
     this.columnConfigurations.push(columnConfiguration);
+    deleteEntry(this.newColumnEditors, editor);
     this.applyChanges();
   }
 
@@ -129,6 +131,12 @@ export class DataCubeExtendManagerState extends DataCubeQuerySnapshotController 
     this.groupExtendedColumns = snapshot.data.groupExtendedColumns.map(
       (col) => new DataCubeQueryExtendedColumnState(col),
     );
+
+    // trigger re-compile in each existing column editor as the base query has changed
+    this.newColumnEditors.forEach((editor) => {
+      editor.getReturnType().catch(noop());
+    });
+    // TODO: existingColumnEditors
   }
 
   applyChanges() {