+++ packages/legend-query-builder/src/stores/fetch-structure/tds/projection/QueryBuilderRelationProjectValueSpecBuidler.ts
@@ -26,6 +26,11 @@ import {
   V1_serializeRawValueSpecification,
   V1_transformRawLambda,
   V1_GraphTransformerContextBuilder,
+  GenericType,
+  GenericTypeExplicitReference,
+  Relation_Relation,
+  Relation_RelationalColumn,
+  Relation_RelationType,
 } from '@finos/legend-graph';
 import { QUERY_BUILDER_SUPPORTED_FUNCTIONS } from '../../../../graph/QueryBuilderMetaModelConst.js';
 import {
@@ -38,7 +43,10 @@ import {
   buildPropertyExpressionChain,
   type LambdaFunctionBuilderOption,
 } from '../../../QueryBuilderValueSpecificationBuilderHelper.js';
-import { UnsupportedOperationError } from '@finos/legend-shared';
+import {
+  guaranteeNonNullable,
+  UnsupportedOperationError,
+} from '@finos/legend-shared';
 
 export const buildRelationProjection = (
   precedingExpression: ValueSpecification,
@@ -46,7 +54,9 @@ export const buildRelationProjection = (
   options?: LambdaFunctionBuilderOption,
 ): SimpleFunctionExpression => {
   const projectFunction = new SimpleFunctionExpression(
-    extractElementNameFromPath(QUERY_BUILDER_SUPPORTED_FUNCTIONS.TDS_PROJECT),
+    extractElementNameFromPath(
+      QUERY_BUILDER_SUPPORTED_FUNCTIONS.RELATION_PROJECT,
+    ),
   );
   const queryBuilderState = tdsState.queryBuilderState;
 
@@ -54,6 +64,7 @@ export const buildRelationProjection = (
 
   const colSepcArray = new ColSpecArray();
   instanceVal.values = [colSepcArray];
+  const relationType = new Relation_RelationType(Relation_RelationType.ID);
 
   tdsState.projectionColumns.forEach((projectionColumnState) => {
     const colSpec = new ColSpec();
@@ -102,7 +113,24 @@ export const buildRelationProjection = (
       );
     }
     colSpec.function1 = columnLambda;
+    const returnType = guaranteeNonNullable(
+      projectionColumnState.getColumnType(),
+      `Can't create value spec for projection column ${projectionColumnState.columnName}`,
+    );
+    relationType.columns.push(
+      new Relation_RelationalColumn(
+        projectionColumnState.columnName,
+        returnType,
+      ),
+    );
   });
   projectFunction.parametersValues = [precedingExpression, instanceVal];
+  const relationGenericType = new GenericType(Relation_Relation.INSTANCE);
+  const relationTypeGenericType = new GenericType(relationType);
+  relationGenericType.typeArguments = [
+    GenericTypeExplicitReference.create(relationTypeGenericType),
+  ];
+  projectFunction.genericType =
+    GenericTypeExplicitReference.create(relationGenericType);
   return projectFunction;
 };