+++ packages/legend-vscode-extension-dependencies/src/index.ts
@@ -17,15 +17,16 @@
 import '../style/index.scss';
 
 export * from './graph-manager/index.js';
+
 export {
-  getDiagram,
-  Point,
-  DIAGRAM_INTERACTION_MODE,
-  DIAGRAM_RELATIONSHIP_EDIT_MODE,
-  DiagramRenderer,
   type Diagram,
   DIAGRAM_ALIGNER_OPERATOR,
+  DIAGRAM_INTERACTION_MODE,
+  DIAGRAM_RELATIONSHIP_EDIT_MODE,
   DIAGRAM_ZOOM_LEVELS,
+  DiagramRenderer,
+  getDiagram,
+  Point,
   V1_diagramModelSchema,
   V1_transformDiagram,
 } from '@finos/legend-extension-dsl-diagram';
@@ -37,27 +38,89 @@ export {
   AlignMiddleIcon,
   AlignStartIcon,
   AlignTopIcon,
+  Button,
   CaretDownIcon,
-  CustomSelectorInput,
-  ControlledDropdownMenu,
+  clsx,
   compareLabelFn,
+  ControlledDropdownMenu,
+  CubesLoadingIndicator,
+  CubesLoadingIndicatorIcon,
+  CustomSelectorInput,
   DistributeHorizontalIcon,
   DistributeVerticalIcon,
+  LegendStyleProvider,
   MenuContent,
   MenuContentDivider,
   MenuContentItem,
-  SaveIcon,
-  clsx,
   MousePointerIcon,
   MoveIcon,
+  SaveCurrIcon,
+  SaveIcon,
+  useResizeDetector,
   ZoomInIcon,
   ZoomOutIcon,
-  useResizeDetector,
-  LegendStyleProvider,
 } from '@finos/legend-art';
 
-export { Class, type PureModel, CORE_PURE_PATH } from '@finos/legend-graph';
+export {
+  type GraphManagerPluginManager,
+  type PureGraphManagerPlugin,
+  type PureGraphPlugin,
+  type PureModel,
+  type PureProtocolProcessorPlugin,
+  Class,
+  Core_GraphManagerPreset,
+  CORE_PURE_PATH,
+  GraphManagerState,
+  Service,
+  V1_PureGraphManager,
+  V1_serializePackageableElement,
+  V1_Service,
+} from '@finos/legend-graph';
 
 export type { Entity } from '@finos/legend-storage';
 
-export type { GeneratorFn } from '@finos/legend-shared';
+export {
+  type GeneratorFn,
+  type PlainObject,
+  AbstractPlugin,
+  AbstractPreset,
+  assertErrorThrown,
+  assertNonNullable,
+  deserializeMap,
+  guaranteeNonEmptyString,
+  guaranteeNonNullable,
+  guaranteeType,
+  isBoolean,
+  SerializationFactory,
+  serializeMap,
+  usingModelSchema,
+} from '@finos/legend-shared';
+
+export {
+  type LegendApplicationConfigurationData,
+  type LegendApplicationConfigurationInput,
+  ApplicationFrameworkProvider,
+  ApplicationStore,
+  ApplicationStoreProvider,
+  BrowserEnvironmentProvider,
+  Core_LegendApplicationPlugin,
+  LegendApplicationConfig,
+  LegendApplicationPlugin,
+  LegendApplicationPluginManager,
+  useApplicationStore,
+} from '@finos/legend-application';
+
+export {
+  type FetchStructureLayoutConfig,
+  type QueryBuilderHeaderActionConfiguration,
+  QueryBuilder_GraphManagerPreset,
+  QueryBuilder_LegendApplicationPlugin,
+  QueryBuilder,
+  QueryBuilderActionConfig,
+  QueryBuilderConfig,
+  QueryBuilderState,
+  QueryBuilderWorkflowState,
+  ServiceQueryBuilderState,
+} from '@finos/legend-query-builder';
+
+export { pureExecution_setFunction } from '@finos/legend-application-studio';