+++ packages/legend-application-pure-ide/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-pure-ide
 
+## 8.0.226
+
+## 8.0.225
+
 ## 8.0.224
 
 ## 8.0.223