+++ packages/legend-application-query/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-query
 
+## 13.5.50
+
+## 13.5.49
+
 ## 13.5.48
 
 ## 13.5.47