+++ packages/legend-application-studio/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-studio
 
+## 28.18.33
+
+## 28.18.32
+
 ## 28.18.31
 
 ## 28.18.30