+++ packages/legend-extension-assortment/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-assortment
 
+## 0.0.201
+
+## 0.0.200
+
 ## 0.0.199
 
 ## 0.0.198