+++ packages/legend-extension-dsl-service/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-service
 
+## 1.0.228
+
+## 1.0.227
+
 ## 1.0.226
 
 ## 1.0.225