+++ packages/legend-extension-dsl-text/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-text
 
+## 6.0.223
+
+## 6.0.222
+
 ## 6.0.221
 
 ## 6.0.220