+++ packages/legend-extension-store-flat-data/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-store-flat-data
 
+## 0.0.295
+
+## 0.0.294
+
 ## 0.0.293
 
 ## 0.0.292